requirements.txt
opencv_python==4.10.0.84
opencv_python_headless==4.9.0.80
Pillow==10.4.0
pyzbar==0.1.9
Requests==2.32.3
